requirements.txt
Django==1.8.6
PyMySQL==0.6.1
django-debug-toolbar==1.4
geopy==0.99
mysql-python==1.2.5
sqlparse==0.1.10
Fabric==1.8.3
django-annoying==0.8.0
django-bootstrap-form==3.1
django-haystack==2.4.1
Whoosh==2.7.0
djangorestframework==3.3.1
djangorestframework-jsonp
django-filter
django-crispy-forms
python-memcached==1.57
requests==2.9.1